Module ironcalc_base::expressions::utils

Functions

column_to_number

Converts column letter identifier to number.

is_valid_column

is_valid_column_number

Checks if column number is in valid range.

is_valid_identifier

is_valid_row

number_to_column

If input number is outside valid range None is returned.

parse_reference_a1

parse_reference_r1c1

quote_name

Quotes a string sheet name if it needs to NOTE: Invalid characters in a sheet name , /, *, [, ], :, ?