Module ironcalc_base::types

Structs

Alignment

Border

BorderItem

CellStyleXfs

CellStyles

CellXfs

Col

Comment

DefinedName

A defined name. The sheet_id is the sheet index in case the name is local

Fill

Font

Metadata

NumFmt

Row

SheetInfo

Information need to show a sheet tab in the UI The color is serialized only if it is not Color::None

Styles

Table

TableColumn

TableStyleInfo

Workbook

An internal representation of an IronCalc Workbook

WorkbookSettings

Worksheet

Internal representation of a worksheet Excel object

Enums

BorderStyle

Cell

CellType

Cell type enum matching Excel TYPE() function values.

FontScheme

HorizontalAlignment

SheetState

Internal representation of a worksheet Excel object

VerticalAlignment

Type Aliases

SheetData

Internal representation of Excel’s sheet_data It is row first and because of this all of our API’s should be row first