Struct ironcalc_base::types::Styles

pub struct Styles {
    pub num_fmts: Vec<NumFmt>,
    pub fonts: Vec<Font>,
    pub fills: Vec<Fill>,
    pub borders: Vec<Border>,
    pub cell_style_xfs: Vec<CellStyleXfs>,
    pub cell_xfs: Vec<CellXfs>,
    pub cell_styles: Vec<CellStyles>,
}

Fields
num_fmts: Vec<NumFmt>fonts: Vec<Font>fills: Vec<Fill>borders: Vec<Border>cell_style_xfs: Vec<CellStyleXfs>cell_xfs: Vec<CellXfs>cell_styles: Vec<CellStyles>
Implementations

impl Styles

pub fn create_new_style(&mut self, style: &Style) -> i32

pub fn get_style_index(&self, style: &Style) -> Option<i32>

pub fn add_named_cell_style( &mut self, style_name: &str, style_index: i32 ) -> Result<(), String>

Adds a named cell style from an existing index Fails if the named style already exists or if there is not a style with that index

pub fn get_style_index_by_name(&self, style_name: &str) -> Result<i32, String>

pub fn create_named_style( &mut self, style_name: &str, style: &Style ) -> Result<(), String>